# [VL]Daily Update Velox Version (20231206)